Moved SliderExt to the testshared module
